```
$ python -m doq.cli --recursive -d /path/to/werkzeug
```

Before
`5.95s user 0.22s system 95% cpu 6.467 total`
After
`5.23s user 0.15s system 96% cpu 5.584 total`